refactoring v1.1
